+++ providers/provider_default_test.go
@@ -2,6 +2,7 @@ package providers
 
 import (
 	"context"
+	"net/url"
 	"testing"
 	"time"
 
@@ -27,6 +28,60 @@ func TestRefresh(t *testing.T) {
 	refreshed, err = p.RefreshSession(context.Background(), nil)
 	assert.False(t, refreshed)
 	assert.Equal(t, ErrNotImplemented, err)
+}
+
+func TestAcrValuesNotConfigured(t *testing.T) {
+	p := &ProviderData{
+		LoginURL: &url.URL{
+			Scheme: "http",
+			Host:   "my.test.idp",
+			Path:   "/oauth/authorize",
+		},
+	}
+
+	result := p.GetLoginURL("https://my.test.app/oauth", "", "", "", "", url.Values{})
+	assert.NotContains(t, result, "acr_values")
+}
+
+func TestAcrValuesConfigured(t *testing.T) {
+	p := &ProviderData{
+		LoginURL: &url.URL{
+			Scheme: "http",
+			Host:   "my.test.idp",
+			Path:   "/oauth/authorize",
+		},
+	}
+
+	result := p.GetLoginURL("https://my.test.app/oauth", "", "", "", "", url.Values{})
+	assert.Contains(t, result, "acr_values=testValue")
+}
+
+func TestCodeChallengeConfigured(t *testing.T) {
+	p := &ProviderData{
+		LoginURL: &url.URL{
+			Scheme: "http",
+			Host:   "my.test.idp",
+			Path:   "/oauth/authorize",
+		},
+	}
+
+	result := p.GetLoginURL("https://my.test.app/oauth", "", "", "challenge", "method", url.Values{})
+	assert.Contains(t, result, "code_challenge=challenge")
+	assert.Contains(t, result, "code_challenge_method=method")
+}
+
+func TestCodeChallengeNotConfigured(t *testing.T) {
+	p := &ProviderData{
+		LoginURL: &url.URL{
+			Scheme: "http",
+			Host:   "my.test.idp",
+			Path:   "/oauth/authorize",
+		},
+	}
+
+	result := p.GetLoginURL("https://my.test.app/oauth", "", "", "", "", url.Values{})
+	assert.NotContains(t, result, "code_challenge")
+	assert.NotContains(t, result, "code_challenge_method")
 }
 
 func TestProviderDataEnrichSession(t *testing.T) {