+++ providers/providers_test.go
@@ -171,3 +171,40 @@ func TestScope(t *testing.T) {
 		g.Expect(pd.Scope).To(Equal(tc.expectedScope))
 	}
 }
+
+func TestForcedMethodS256(t *testing.T) {
+	g := NewWithT(t)
+	options := options.NewOptions()
+	options.Providers[0].ForceCodeChallengeMethod = CodeChallengeMethodS256
+	method := parseCodeChallengeMethod(options.Providers[0])
+
+	g.Expect(method).To(Equal(CodeChallengeMethodS256))
+}
+
+func TestForcedMethodPlain(t *testing.T) {
+	g := NewWithT(t)
+	options := options.NewOptions()
+	options.Providers[0].ForceCodeChallengeMethod = CodeChallengeMethodPlain
+	method := parseCodeChallengeMethod(options.Providers[0])
+
+	g.Expect(method).To(Equal(CodeChallengeMethodPlain))
+}
+
+func TestPrefersS256(t *testing.T) {
+	g := NewWithT(t)
+	options := options.NewOptions()
+	options.Providers[0].CodeChallengeMethods = []string{"plain", "S256"}
+	method := parseCodeChallengeMethod(options.Providers[0])
+
+	g.Expect(method).To(Equal(CodeChallengeMethodS256))
+}
+
+func TestCanOverwriteS256(t *testing.T) {
+	g := NewWithT(t)
+	options := options.NewOptions()
+	options.Providers[0].CodeChallengeMethods = []string{"plain", "S256"}
+	options.Providers[0].ForceCodeChallengeMethod = "plain"
+	method := parseCodeChallengeMethod(options.Providers[0])
+
+	g.Expect(method).To(Equal(CodeChallengeMethodPlain))
+}